# Conflicts:
#	packages/wallet/wallet-toolbox/src/storage/methods/reviewStatusIdb.ts

+++ packages/overlays/topics/src/__tests__/kvstore.test.ts
@@ -258,9 +258,9 @@ describe('KVStoreTopicManager', () => {
 // ---------------------------------------------------------------------------
 
 describe('KVStoreLookupService (MongoDB)', () => {
-  let mongod: MongoMemoryServer
-  let client: MongoClient
-  let db: Db
+  let mongod: MongoMemoryServer | undefined
+  let client: MongoClient | undefined
+  let db: Db | undefined
   let service: ReturnType<typeof KVStoreLookupServiceFactory>
 
   beforeAll(async () => {
@@ -273,12 +273,18 @@ describe('KVStoreLookupService (MongoDB)', () => {
   })
 
   afterAll(async () => {
-    await client.close()
-    await mongod.stop()
+    if (client !== undefined) {
+      await client.close()
+    }
+    if (mongod !== undefined) {
+      await mongod.stop()
+    }
   })
 
   afterEach(async () => {
-    await db.collection('kvstoreRecords').deleteMany({})
+    if (db !== undefined) {
+      await db.collection('kvstoreRecords').deleteMany({})
+    }
   })
 
   /**
